[upd] ArticlesPage - separate article list to ArticleInfiniteList component
