+++ src/pages/ArticlesPage/ui/ArticlesPage/ArticlesPage.tsx
@@ -1,18 +1,13 @@
 import { type FC, memo, useCallback } from 'react';
-import { useSelector } from 'react-redux';
-import { ArticleList } from 'entities/Article';
 import { classNames } from 'shared/lib/classNames/classNames';
-import { useAppDispatch } from 'shared/lib/hooks/useAppDispatch/useAppDispatch';
-import { useInitialEffect } from 'shared/lib/hooks/useInitialEffect/useInitialEffect';
 import { DynamicModuleLoader, type ReducerList } from 'shared/lib/components/DynamicModuleLoader';
-import { articlesPageReducer, getArticles } from '../../model/slice/articlesPageSlice';
-import { getArticlesPageIsLoading, getArticlesPageView } from '../../model/selectors/articlesPage';
+import { articlesPageReducer } from '../../model/slice/articlesPageSlice';
 import cls from './ArticlesPage.module.scss';
-import { fetchNextArticles } from '../../model/services/fetchNextArticles/fetchNextArticles';
-import { initArticlesPage } from '../../model/services/initArticlesPage/initArticlesPage';
 import { Page } from 'widgets/Page';
 import { ArticlesPageFilter } from '../ArticlesPageFilter/ArticlesPageFilter';
-import { useSearchParams } from 'react-router-dom';
+import { ArticleInfiniteList } from '../../ui/ArticleInfiniteList/ArticleInfiniteList';
+import { fetchNextArticles } from '../../model/services/fetchNextArticles/fetchNextArticles';
+import { useAppDispatch } from 'shared/lib/hooks/useAppDispatch/useAppDispatch';
 
 interface ArticlesPageProps {
     className?: string;
@@ -26,19 +21,11 @@ export const ArticlesPage: FC<ArticlesPageProps> = (props) => {
     const {
         className = ''
     } = props;
-    const [searchParams] = useSearchParams();
     const dispatch = useAppDispatch();
-    const articles = useSelector(getArticles.selectAll);
-    const view = useSelector(getArticlesPageView);
-    const isLoading = useSelector(getArticlesPageIsLoading);
 
     const onLoadNextPart = useCallback(() => {
         dispatch(fetchNextArticles());
     }, [dispatch]);
-
-    useInitialEffect(() => {
-        dispatch(initArticlesPage(searchParams));
-    });
 
     return (
         <DynamicModuleLoader reducers={ reducers } removeAfterUnmount={ false }>
@@ -47,12 +34,7 @@ export const ArticlesPage: FC<ArticlesPageProps> = (props) => {
                 className={ classNames(cls.ArticlesPage, {}, [className]) }
             >
                 <ArticlesPageFilter/>
-                <ArticleList
-                    className={ cls.list }
-                    isLoading={ isLoading }
-                    articleList={ articles }
-                    view={ view }
-                />
+                <ArticleInfiniteList/>
             </Page>
         </DynamicModuleLoader>
